requirements.txt
discord.py
python-dotenv
scipy
numpy
pytest
python-dateutil